requirements.txt
absl-py==0.13.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.3.0
altair==4.1.0
anyio==2.2.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
atari-py==0.2.9
attrs==20.3.0
babel==2.9.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.3
biopython==1.78
bleach==3.3.0
boto3==1.17.28
botocore==1.20.28
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.2
cryptography==3.2.1
cycler==0.10.0
datasets==1.9.0
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
editdistance==0.5.3
emcee==3.0.2
emukit==0.4.8
entrypoints==0.3
fair-esm==0.3.1
ffmpeg==1.4
filelock==3.0.12
fire==0.4.0
flatbuffers==2.0
fsspec==2021.6.1
gast==0.4.0
gitdb==4.0.7
gitpython==3.1.18
globus-cli==2.0.0
google-auth-oauthlib==0.4.5
google-auth==1.34.0
google==3.0.0
gpy==1.9.9
grpcio==1.34.1
gym==0.18.3
h5py==3.2.1
hmmlearn==0.2.5
huggingface-hub==0.0.12
idna==2.10
imagecorruptions==1.1.2
imageio==2.9.0
importlib-metadata==3.7.2
ipykernel==5.5.0
ipython-genutils==0.2.0
ipython==7.21.0
ipywidgets==7.6.3
jedi==0.18.0
jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
jupyterlab==3.0.10
keras-nightly==2.5.0.dev2021032900
kiwisolver==1.3.1
lmdb==1.1.1
markdown==3.3.4
markupsafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
multipledispatch==0.6.0
multiprocess==0.70.12.2
natsort==7.1.1
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.2.0
numexpr==2.7.3
numpy==1.19.2
oauthlib==3.1.1
olefile==0.46
olive-oil-ml==0.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.2
paramz==0.9.5
parso==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.1.2
pip==20.2.4
plotly==4.14.3
portpicker==1.4.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==4.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycosat==0.6.3
pycparser==2.20
pydoe==0.3.8
pyglet==1.5.15
pygments==2.8.1
pyopenssl==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pysocks==1.7.1
python-dateutil==2.8.1
python-levenshtein==0.12.2
pytz==2021.1
pywavelets==1.1.1
pyyaml==5.4.1
pyzmq==19.0.2
regex==2020.11.13
requests-oauthlib==1.3.0
requests==2.24.0
retrying==1.3.3
rsa==4.7.2
ruamel-yaml==0.15.87
s3transfer==0.3.4
sacremoses==0.0.45
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
selenium==3.141.0
send2trash==1.5.0
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.2.1
stable-baselines3==1.1.0
subprocess32==3.5.4
tables==3.6.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.17
tokenizers==0.10.3
toolz==0.11.1
torch==1.9.0
torchcontrib==0.0.2
torchdiffeq==0.2.2
torchvision==0.10.0
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
transformers==4.8.2
typing-extensions==3.7.4.3
urllib3==1.25.11
wandb==0.10.28
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.35.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.1
xxhash==2.0.2
zipp==3.4.1